modified the convert video script - only converts when the output file does not already exist

+++ h264_bag_playback/scripts/convert_video.sh
@@ -9,9 +9,14 @@ find $input_folder/* -prune -type d | while IFS= read -r d; do
       output_file=$(echo $input_file | sed "s/.h264/.mp4/g")
       temp_file=$(echo $input_file | sed "s/.h264/.tmp.mp4/g")
       echo "Converting $input_file to $output_file"
-      ffmpeg -y -i $input_file -map 0:v -vcodec copy -bsf:v h264_mp4toannexb $temp_file
-      ffmpeg -y -fflags +genpts -r 30 -i $temp_file -vcodec copy $output_file
-      rm $temp_file
+      if [ ! -f "$output_file" ]; then
+        echo "Will convert $input_file"
+        ffmpeg -y -i $input_file -map 0:v -vcodec copy -bsf:v h264_mp4toannexb $temp_file
+        ffmpeg -y -fflags +genpts -r 30 -i $temp_file -vcodec copy $output_file
+        rm $temp_file
+      else
+        echo "$output_file already exists"
+      fi
     fi
   done
 done
